refactor: resolvendo algumas pendências que ficaram
